pyproject.toml
[project]
name = "mcp-md"
version = "0.1.0"
description = "A simple LLM Agent with MCP support"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.1.0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]


[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64"]

[tool.pixi.pypi-dependencies]
mcp-md = { path = ".", editable = true }

[tool.pixi.environments]
default = { solve-group = "default" }
dev = { features = ["dev"], solve-group = "default" }

[tool.pixi.tasks]
dev = "python src/test.py"
start = "python src/main.py"
test-mcp = "python test_mcp.py"
lint = "ruff check src/"
format = "ruff format src/"
